+++ database/migrations/core/2025_12_29_200807_create_personal_access_tokens_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::getConnection()->getName() !== 'pgsql') {
+            return;
+        }
+
         Schema::create('personal_access_tokens', function (Blueprint $table) {
             $table->id();
             $table->morphs('tokenable');
@@ -28,6 +32,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (Schema::getConnection()->getName() !== 'pgsql') {
+            return;
+        }
+
         Schema::dropIfExists('personal_access_tokens');
     }
 };
